copy actions save button above actiosn list
I created just a handful of actions, and then later in the day came back
to create another and for the life of me I could not figure out how to
add a new action. Could not find the button. So maybe if there are more
than a few, there should be another copy of the button above the list.

Confession: I had difficulty getting the app to run locally so I'm not
sure if this works.