Fix failing settings_controller_spec
It's failing due to an error raised by Stripe, which is that the user
doesn't have a customer_id, so it can't look them up. Giving the user a
customer id leads to configuration errors, because I don't have any
stripe credentials and it's trying to talk to stripe.

To avoid this, I stubbed some methods and described my expectations
about which would be called. To make this easier, I created a Customer
class, which wraps an instance of Stripe::Customer and has a more
concise external API.

The controller will still fail when a user doesn't have a customer_id
and tries to update their credit card, but I don't know if that's
possible?

hopefully resolves #91