Fix a stylesheet typo
I happened to inspect element and noticed this rule:

```
overflow-wrap: break-word strong;
```

Firefox shows a little warning icon and says that "break-word strong" is
an "invalid property value". I think this looks like it's a typo, and I
presume this is what it meant to say. Not sure how to test what this
actually looks like, but figured I'd toss the PR over the fence for your
consideration.